Run differential expression analysis

Run differential expression analysis

run_dea(sample.info, reference = NULL, species = c("human", "mouse"),
  output.dir, threads = NULL, fastqc = "fastqc", multiqc = "multiqc",
  index.dir = NULL, salmon = "salmon", bam = FALSE, bootstraps = 0,
  seqBias = TRUE, gcBias = TRUE, posBias = FALSE,
  allowOrphans = FALSE, advanced.opts = NULL)

Arguments

sample.info
character string giving the path to a tab-delimited text file with at least the columns <condition> (treatment condition), <sample> (sample name), and <file1> (absolute or relative path to the fastq or salmon quant.sf files). If fastq files and PE reads, then a column <file2> should also be present. If a batch effect is to be included in the design, then this should be identified under the column <batch>.

reference
character vector specifying the conditions in order. For example, c("A", "B", "C", "D") would mean "A" is the reference condition to which "B", "C" and "D" are compared; in addition, "C" and "D" will be compared to "B", and "D" will be compared to "C". If NULL then the comparisons will be arranged alphabetically. [DEFAULT = NULL].

species
character string specifying the name of the species. Only 'human', and 'mouse' are supported at present. [DEFAULT = human].

output.dir
character string specifying the directory to which results will be saved. If the directory does not exist, it will be created.

threads
an integer value indicating the number of parallel threads to be used by FastQC. [DEFAULT = maximum number of available threads - 1].

fastqc
a character string specifying the path to the fastqc executable. [DEFAULT = "fastqc"].

multiqc
a character string specifying the path to the multiqc executable. [DEFAULT = "multiqc"].

index.dir
directory of the index files needed for read mapping using Salmon. See function 'build_index()'.

salmon
a character string specifying the path to the salmon executable. [DEFAULT = "salmon"].

bam
logical, if TRUE then create a pseudo-alignment BAM file. [Default = FALSE]

bootstraps
integer giving the number of bootstrap samples that Salmon should use (default is 0). With bootstrap samples, uncertainty in abundance can be quantified.

seqBias
logical, should Salmon's option be used to model and correct abundances for sequence specific bias? Default is TRUE.

gcBias
logical, should Salmon's option be used to model and correct abundances for GC content bias? Requires Salmon version 0.7.2 or higher. Default is TRUE.

posBias
logical, should Salmon's option be used to model and correct abundances for positional biases? Requires Salmon version 0.7.3 or higher. Default is FALSE.

allowOrphans
logical, if TRUE then consider orphaned reads as valid hits when performing lightweight-alignment. This option will increase sensitivity (allow more reads to map and more transcripts to be detected), but may decrease specificity as orphaned alignments are more likely to be spurious. For more details see Salmon documentation.

advanced.opts
character vector supplying list of advanced option arguments to apply to each Salmon call. For details see Salmon documentation or type salmon quant --help-reads at the command line.

Examples

# NOT RUN {
targets.file <- data.frame(
  file1 = list.files(system.file("extdata/FASTQ/", package = "rseqR"),
                     "*_1.fastq", full.names = TRUE),
  file2 = list.files(system.file("extdata/FASTQ/", package = "rseqR"),
                     "*_2.fastq", full.names = TRUE),
  sample = c("BUFF1", "BUFF2", "OX1", "OX2", "UNT1", "UNT2"),
  condition = c(rep("BUFF", 2), rep("OX", 2), rep("UNT", 2)),
  batch = rep(1:2, times = 3))
write.table(targets.file, "targets.txt", col.names = TRUE,
  row.names = FALSE, sep = "\t", quote = FALSE)
run_dea(sample.info = "targets.txt", reference = c("UNT", "BUFF", "OX"),
        species = "human", output.dir = "results", index.dir = NULL)
# }