Wrapper function to run tximport

run_tximport function runs tximport on transcript level abundances from Salmon to summarise to gene level. See Bioconductor package tximport for details.

run_tximport(species = c("human", "mouse"), salmon.files,
  summarise = TRUE, countsFromAbundance = c("no", "scaledTPM",
  "lengthScaledTPM"))

Arguments

species
character string specifying the name of the species. Only 'human', and 'mouse' are supported at present. [DEFAULT = human].

salmon.files
character vector specifying path to quant.sf files.

summarise
logical, if TRUE then return gene-level summary, otherwise return transcript level data.

countsFromAbundance
whether to generate counts based on abundance. Available options are: 'no', 'scaledTPM' (abundance based estimated counts scaled up to library size), 'lengthScaledTPM'.

Value

a list of gene and transcript level estimated counts.

Details

We use Ensembl annotation for both genes and transcripts.

References

Charlotte Soneson, Michael I. Love, Mark D. Robinson (2015): Differential analyses for RNA-seq: transcript-level estimates improve gene-level inferences. F1000Research. http://dx.doi.org/10.12688/f1000research.7563.1